feat(sidebar): resume last decopilot thread on Home click
Lift useNavigateToAgentThread into a shared hook and use it for both
the pinned-agent buttons and the sidebar Home button. Clicking Home now
resumes the user's most recent decopilot thread when one exists in the
local cache, falling back to creating a fresh decopilot task otherwise.
The /\$org route-level redirect is unchanged, so direct URL navigation
still mints a new task.
