+++ apps/mesh/src/web/hooks/use-project-sidebar-items.tsx
@@ -1,4 +1,4 @@
-import { useProjectContext } from "@decocms/mesh-sdk";
+import { getDecopilotId, useProjectContext } from "@decocms/mesh-sdk";
 import type {
   NavigationSidebarItem,
   SidebarSection,
@@ -15,6 +15,7 @@ import { usePanelActions } from "@/web/layouts/shell-layout";
 import { formatPinnedViewTabId } from "@/web/layouts/main-panel-tabs/tab-id";
 import { pluginRootSidebarItems, pluginSidebarGroups } from "../index.tsx";
 import { PLUGIN_ID as WORKFLOWS_PLUGIN_ID } from "mesh-plugin-workflows/shared";
+import { useNavigateToAgentThread } from "@/web/hooks/use-navigate-to-agent-thread";
 
 export function useProjectSidebarItems(): SidebarSection[] {
   const { org: orgContext } = useProjectContext();
@@ -23,6 +24,7 @@ export function useProjectSidebarItems(): SidebarSection[] {
   const { setTasksOpen, openTab } = usePanelActions();
   const org = orgContext.slug;
   const currentProject = useProjectContext().project;
+  const navigateToAgentThread = useNavigateToAgentThread(org);
 
   const routeParams = useParams({ strict: false }) as {
     taskId?: string;
@@ -157,7 +159,7 @@ export function useProjectSidebarItems(): SidebarSection[] {
     isActive: pathname === `/${org}` || pathname === `/${org}/`,
     onClick: () => {
       setTasksOpen(false);
-      navigate({ to: "/$org", params: { org } });
+      void navigateToAgentThread(getDecopilotId(orgContext.id));
     },
   };
